+++ dbfread/__init__.py
@@ -6,18 +6,16 @@ __url__ = 'http://nerdly.info/ole/'
 __license__ = 'MIT'
 __version__ = '1.0.6'
 
-from .dbf import Table
-from .dbf import Table as open  # Alias.
+from .dbf import DBF
+from .dbf import DBF as open  # Alias.
 from . import dbf as _dbf
 
 def read(filename, load=True, **kwargs):
-    """Read a DBF file and return a list of records.
-
-    Returns a OldStyleTable object which is also a list of records.
+    """Read a DBF file and return a DeprecatedDBF object.
 
     This function is deprecated. Please use open(load=True) instead.
     """
     from warnings import warn
     warn("read() is deprecated. Please use dbfread.open(filename, load=True)")
 
-    return _dbf.OldStyleTable(filename, load=load, **kwargs)
+    return _dbf.DeprecatedDBF(filename, load=load, **kwargs)